refactor: 분리된 state, action 별도 파일로 분리
